Fix sbt plugin artifact ids
Contrary to what's in the pom, the artifactId of an sbt plugin is
"$artifactName$binaryVersion" otherwise it is not unique
and the MavenReference is wrong.

This fix sbt plugins to be constantly re-indexed by the SonatypeSynchronizer
